+++ 1 The Physical/Guess Point Level 1 (Stateless)/2 - Arrange-Act-Assert Backwards/2.1 - Password Validator/src/password-validator.ts
@@ -1,7 +1,6 @@
 import { Errors } from "./errors";
 
 export type PasswordValidation = { status: boolean, errors: Array<Errors> };
-
 export class PasswordValidator {
     validate(password: string): PasswordValidation {
         const passwordValidation: PasswordValidation = {
@@ -14,19 +13,19 @@ export class PasswordValidator {
             passwordValidation.errors.push(Errors.NO_DIGITS);
         }
 
-        if (this.doesNotMaxLength(password)) {
-            passwordValidation.status = false;
-            passwordValidation.errors.push(Errors.TOO_LONG);
-        }
-
-        if (this.doesNotHaveMinLength(password)) {
-            passwordValidation.status = false;
-            passwordValidation.errors.push(Errors.TOO_SHORT);
-        }
-
         if (this.doesNotHaveUppercase(password)) {
             passwordValidation.status = false;
             passwordValidation.errors.push(Errors.NO_UPPER);
+        }
+
+        if (this.isOverMaxLength(password)) {
+            passwordValidation.status = false;
+            passwordValidation.errors.push(Errors.TOO_LONG);
+
+        }
+        if (this.isUnderMinLength(password)) {
+            passwordValidation.status = false;
+            passwordValidation.errors.push(Errors.TOO_SHORT);
         }
 
         return passwordValidation;
@@ -35,11 +34,11 @@ export class PasswordValidator {
         const uppercasePatternRegex: RegExp = /.*[A-Z].*/g;
         return password.match(uppercasePatternRegex) === null;
     }
-    private doesNotHaveMinLength(password: string):boolean {
+    private isUnderMinLength(password: string):boolean {
         const MIN_LENGTH_FOR_A_PASSWORD = 5;
         return password.length < MIN_LENGTH_FOR_A_PASSWORD;
     }
-    private doesNotMaxLength(password: string): boolean {
+    private isOverMaxLength(password: string): boolean {
         const MAX_LENGTH_FOR_A_PASSWORD = 15;
         return password.length > MAX_LENGTH_FOR_A_PASSWORD;
     }
